+++ site/DFM.Entities/Month.cs
@@ -11,7 +11,7 @@ namespace DFM.Entities
 			init();
 		}
 
-		public virtual Int32 ID { get; set; }
+		public virtual Int64 ID { get; set; }
 
 		public virtual Int16 Time { get; set; }
